Create_stat_for_thickness_of_50_brain_regions
Contained within this file is the comprehensive code designed to calculate essential statistical metrics, including mean, standard deviation (SD), p-values, and T-values, specifically focusing on the thickness of various brain regions. The intention behind this code is to provide a straightforward and user-friendly solution, streamlining the process for users.

To utilize this code effectively, you only need to incorporate a CSV file that encompasses the necessary data on all the brain regions of interest. This simple integration allows for seamless functionality, ensuring that users, even those with minimal coding expertise, can easily apply the code to their datasets. Our emphasis on user-friendliness aims to facilitate a smoother and more efficient experience in deriving meaningful insights from the thickness data of diverse brain regions.